Initial commit
Commits the framework for creating the cli apps. Also attempts to
imagine what the printing of the help would look like.

The only personal note for the use of OptionFunc is that the clients
functions that are provided that satisfy the type must handle the
arguments that are passed in themselves